Update expander graph generation code
Updated expander graph generation code for better compatability with GNN
code. Now explicitly add the expander graph nodes to the original graph
structure, and add a mask, where mask[i] == 1 if node i in the graph
belongs to the original graph, and 0 if it is a node inserted as part of
the expander graph.