Stop propagation on expander graph in final layer
Previously we performed propagation on the expander graph in every
layer. However, the original expander graph propagation paper only
performs message passing on the expander graph in even layers. This
update makes sure that the final layer performs message passing on the
original graph, and not the expander graph.

For instance, if we set `num_layer` = 3 then we will perform 3 rounds of
message passing on the original graph, and 2 on the expander graph,
alternating between original and expander graph.